Refactoring session. separated create article to new component. articles.js only takes 'articles' as argument instead of category, title and body.
